+++ relay-server/src/actors/processor.rs
@@ -1145,11 +1145,7 @@ impl EnvelopeProcessorService {
                             item.set_payload(ContentType::OctetStream, recording.as_slice());
                         }
                         Err(e) => {
-                            relay_log::warn!(
-                                "replay-recording-event: failed to parse {:?} with message {}",
-                                event_id,
-                                e
-                            );
+                            relay_log::warn!("replay-recording-event: {e} {event_id:?}");
                             context.track_outcome(
                                 Outcome::Invalid(DiscardReason::InvalidReplayRecordingEvent),
                                 DataCategory::Replay,